add storage logic

+++ PetsFactsApp/Models/CoreDataModels/CatFact+CoreDataProperties.swift
@@ -10,12 +10,11 @@ import Foundation
 import CoreData
 
 @objc(CatFact)
-public class CatFact: NSManagedObject{
-}
+public class CatFact: NSManagedObject {}
 
 extension CatFact {
     @NSManaged public var name: String?
     @NSManaged public var fact: String
     @NSManaged public var color: String
-    @NSManaged public var likes: Int
+    @NSManaged public var likes: Int64
 }
